+++ module/actor/actor-beast-sheet.js
@@ -137,9 +137,16 @@ export class ArM5eBeastActorSheet extends ArM5eActorSheet {
         log(false, "Invalid drop");
         return false;
       }
+    } else if (type == "ability") {
+      if (this.actor.hasSkill(itemData.data.key)) {
+        ui.notifications.warn(`This character already have the ability: ${itemData.name}`);
+      }
     }
     // }
     const res = await super._onDropItem(event, data);
+    if (res.length == 1) {
+      res[0].sheet.render(true);
+    }
     return res;
   }
 }